moved epoch default value to cdf module
