+++ imap_processing/mag/l0/decom_mag.py
@@ -11,8 +11,9 @@ import numpy as np
 import xarray as xr
 from space_packet_parser import parser, xtcedef
 
-from imap_processing import epoch_attrs, imap_module_directory
+from imap_processing import imap_module_directory
 from imap_processing.ccsds.ccsds_data import CcsdsData
+from imap_processing.cdf import epoch_attrs
 from imap_processing.cdf.imap_cdf_manager import ImapCdfAttributes
 from imap_processing.cdf.utils import met_to_j2000ns
 from imap_processing.mag.constants import DataMode
