+++ imap_processing/mag/l1a/mag_l1a.py
@@ -7,7 +7,7 @@ from pathlib import Path
 import numpy as np
 import xarray as xr
 
-from imap_processing import epoch_attrs
+from imap_processing.cdf import epoch_attrs
 from imap_processing.cdf.imap_cdf_manager import ImapCdfAttributes
 from imap_processing.cdf.utils import J2000_EPOCH, met_to_j2000ns
 from imap_processing.mag.constants import DataMode, PrimarySensor
