+++ imap_processing/utils.py
@@ -10,7 +10,7 @@ import pandas as pd
 import xarray as xr
 from space_packet_parser import parser, xtcedef
 
-from imap_processing import epoch_attrs
+from imap_processing.cdf import epoch_attrs
 from imap_processing.cdf.imap_cdf_manager import ImapCdfAttributes
 from imap_processing.cdf.utils import met_to_j2000ns
